% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.psychnet_bootstrap}
\alias{print.psychnet_bootstrap}
\title{Print a network bootstrap}
\usage{
\method{print}{psychnet_bootstrap}(x, ...)
}
\arguments{
\item{x}{A \code{psychnet_bootstrap} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a network bootstrap
}
