% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct.R
\name{print.psychnet_nct}
\alias{print.psychnet_nct}
\title{Print a Network Comparison Test}
\usage{
\method{print}{psychnet_nct}(x, ...)
}
\arguments{
\item{x}{A \code{psychnet_nct} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Network Comparison Test
}
